pyproject.toml
[project]
name = "advanced-ai-documentation-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "fastapi>=0.122.0",
    "langchain>=1.0.7",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
]
